refactoring: moved 'commits_to_file' to 'Project' class
